app/requirements.txt
flask
requests
python-dotenv
cachetools